Do not call `shm_unlink` just after we launch the game server and the player's program
This is because that:

1) Racing condition. We do not know whether the `shm_unlink` is called before
or after the game server and the player's program opens the shm region. If it
is called before, the game server and the player's program will be unable to
open the shm region, leading to something like "killed by signal 7 (SIGBUS)"

2) We can call `shm_unlink` just before our judger exits. This does not work
when the judger crashes, but... we just simply ignore that possibility...

I'm too tired and busy. I have no idea why we're still preparing and debugging
problems on one day before the HPCGame competition! It's really weird. The
timetable is so compact, which I think will affect the quality of problems
seriously...